requirements.txt
Django==1.10
django-paypal==0.3.2
Pillow==3.3.1
pytz==2016.6.1
requests==2.11.1
six==1.10.0
